added 2 new parameters to ofxKinectNui::init() function: 'grabsCalibrationVideo' for point cloud, 'grabsLabelCv' for using label data with openCV. added example with openCV: find contour from depth pixels data with near/ far clipping. find contour from label data.

+++ example/src/testApp.h
@@ -69,8 +69,8 @@ class testApp : public ofBaseApp {
 		bool bPlugged;
 		bool bUnplugged;
 		
-		float nearClipping;
-		float farClipping;
+		unsigned short nearClipping;
+		unsigned short farClipping;
 		int angle;
 		
 		int mRotationX, mRotationY;
